Patient diagnosis and Demographic info improvement
Added schemas for both these dataframe, in order to  avoid problems with InferSchema which doesn't infer the right types. Doing so with Patient diagnosis no problem, instead Demographic Info comes out with strange behavior...I skipped It using the Tapullata Increddibbile; in the future we need to correct it. The function calculating the BMI is now redesigned.